项目2次

+++ app/src/main/java/com/example/xiangmu/view/ShowBean.java
@@ -411,7 +411,7 @@ public class ShowBean {
             private String pic_url;
             private int sort_order;
             private int is_show;
-            private int floor_price;
+            private double floor_price;
             private String app_list_pic_url;
             private int is_new;
             private String new_pic_url;
@@ -473,7 +473,7 @@ public class ShowBean {
                 this.is_show = is_show;
             }
 
-            public int getFloor_price() {
+            public double getFloor_price() {
                 return floor_price;
             }
 
@@ -522,7 +522,7 @@ public class ShowBean {
             private String item_pic_url;
             private String subtitle;
             private int topic_category_id;
-            private int price_info;
+            private double price_info;
             private String read_count;
             private String scene_pic_url;
             private int topic_template_id;
@@ -586,7 +586,7 @@ public class ShowBean {
                 this.topic_category_id = topic_category_id;
             }
 
-            public int getPrice_info() {
+            public double getPrice_info() {
                 return price_info;
             }
 
@@ -683,7 +683,7 @@ public class ShowBean {
                 private int id;
                 private String name;
                 private String list_pic_url;
-                private int retail_price;
+                private double retail_price;
 
                 public int getId() {
                     return id;
@@ -709,7 +709,7 @@ public class ShowBean {
                     this.list_pic_url = list_pic_url;
                 }
 
-                public int getRetail_price() {
+                public double getRetail_price() {
                     return retail_price;
                 }
